+++ aiohttp/helpers.py
@@ -15,6 +15,8 @@ from urllib.parse import quote, urlencode
 
 from multidict import MultiDict, MultiDictProxy
 
+from async_timeout import timeout
+
 from . import hdrs
 from .errors import InvalidURL
 
@@ -29,6 +31,7 @@ __all__ = ('BasicAuth', 'create_future', 'FormData', 'parse_mimetype',
 
 
 sentinel = object()
+Timeout = timeout
 
 
 class BasicAuth(namedtuple('BasicAuth', ['login', 'password', 'encoding'])):
@@ -519,53 +522,6 @@ def is_ip_address(host):
     else:
         raise TypeError("{} [{}] is not a str or bytes"
                         .format(host, type(host)))
-
-
-class Timeout:
-    """Timeout context manager.
-
-    Useful in cases when you want to apply timeout logic around block
-    of code or in cases when asyncio.wait_for is not suitable. For example:
-
-    >>> with aiohttp.Timeout(0.001):
-    ...     async with aiohttp.get('https://github.com') as r:
-    ...         await r.text()
-
-
-    :param timeout: timeout value in seconds or None to disable timeout logic
-    :param loop: asyncio compatible event loop
-    """
-    def __init__(self, timeout, *, loop=None):
-        self._timeout = timeout
-        if loop is None:
-            loop = asyncio.get_event_loop()
-        self._loop = loop
-        self._task = None
-        self._cancelled = False
-        self._cancel_handler = None
-
-    def __enter__(self):
-        self._task = asyncio.Task.current_task(loop=self._loop)
-        if self._task is None:
-            raise RuntimeError('Timeout context manager should be used '
-                               'inside a task')
-        if self._timeout is not None:
-            self._cancel_handler = self._loop.call_later(
-                self._timeout, self._cancel_task)
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if exc_type is asyncio.CancelledError and self._cancelled:
-            self._cancel_handler = None
-            self._task = None
-            raise asyncio.TimeoutError from None
-        if self._timeout is not None:
-            self._cancel_handler.cancel()
-            self._cancel_handler = None
-        self._task = None
-
-    def _cancel_task(self):
-        self._cancelled = self._task.cancel()
 
 
 def _get_kwarg(kwargs, old, new, value):